chore: rename cc to ophelia
Resolve crate name conflict
