--- cc-hasher-blake2b/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "cc-hasher-blake2b"
-version = "0.1.0"
-authors = ["zero.qn <zero.qn@outlook.com>"]
-edition = "2018"
-license = "MIT"
-
-[dependencies]
-cc-hasher = { path = "../cc-hasher" }
-
-blake2b_simd = "0.5"
-
-[dev-dependencies]
-quickcheck = "0.8"
-quickcheck_macros = "0.8"
