--- cc-secp256k1/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "cc-secp256k1"
-version = "0.1.0"
-authors = ["zero.qn <zero.qn@outlook.com>"]
-edition = "2018"
-license = "MIT"
-
-[dependencies]
-cc = { path = "../cc" }
-cc-derive = { path = "../cc-derive" }
-
-secp256k1 = { version = "0.14", features = [ "rand" ] }
-lazy_static = "1.3"
-rand = { version = "0.6", optional = true }
-
-[dev-dependencies]
-cc = { path = "../cc", features = [ "proptest" ] }
-
-rand = "0.6"
-sha2 = "0.8"
-quickcheck = "0.8"
-quickcheck_macros = "0.8"
-
-[features]
-default = []
-generate = [ "cc/generate", "rand" ]
