--- cc/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "cc"
-version = "0.1.0"
-authors = ["zero.qn <zero.qn@outlook.com>"]
-edition = "2018"
-license = "MIT"
-
-[dependencies]
-cc-hasher = { path = "../cc-hasher" }
-
-[dependencies.rand]
-version = "0.6"
-optional = true
-default-features = false
-
-[dependencies.quickcheck]
-version = "0.8"
-optional = true
-default-features = false
-
-[dependencies.cc-quickcheck-types]
-path = "../cc-quickcheck-types"
-optional = true
-default-features = false
-
-[dev-dependencies]
-quickcheck = "0.8"
-quickcheck_macros = "0.8"
-
-[features]
-default = []
-generate = [ "rand" ]
-proptest = [ "quickcheck", "cc-quickcheck-types", "cc-hasher/proptest" ]
