+++ ophelia-derive/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "ophelia-derive"
+version = "0.1.0"
+authors = ["zero.qn <zero.qn@outlook.com>"]
+edition = "2018"
+license = "MIT"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = "0.15"
+quote = "0.6"
