+++ ophelia-hasher-sm3/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "ophelia-hasher-sm3"
+version = "0.1.0"
+authors = ["zero.qn <zero.qn@outlook.com>"]
+edition = "2018"
+
+[dependencies]
+ophelia-hasher = { path = "../ophelia-hasher" }
+
+libsm = { git = "https://github.com/cryptape/libsm", rev = "ac323abd3512a9fdb8bfd6cd349a6fc46deb688f" }
+
+[dev-dependencies]
+quickcheck = "0.8"
+quickcheck_macros = "0.8"
