+++ ophelia-hasher/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "ophelia-hasher"
+version = "0.1.0"
+authors = ["zero.qn <zero.qn@outlook.com>"]
+edition = "2018"
+license = "MIT"
+
+[dependencies]
+quickcheck = { version = "0.8", default-features = false, optional = true }
+
+[dev-dependencies]
+quickcheck = "0.8"
+quickcheck_macros = "0.8"
+
+[features]
+proptest = [ "quickcheck" ]
