+++ ophelia-secp256k1/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "ophelia-secp256k1"
+version = "0.1.0"
+authors = ["zero.qn <zero.qn@outlook.com>"]
+edition = "2018"
+license = "MIT"
+
+[dependencies]
+ophelia = { path = "../ophelia" }
+ophelia-derive = { path = "../ophelia-derive" }
+
+secp256k1 = { version = "0.14", features = [ "rand" ] }
+lazy_static = "1.3"
+rand = { version = "0.6", optional = true }
+
+[dev-dependencies]
+ophelia = { path = "../ophelia", features = [ "proptest" ] }
+
+rand = "0.6"
+sha2 = "0.8"
+quickcheck = "0.8"
+quickcheck_macros = "0.8"
+
+[features]
+default = []
+generate = [ "ophelia/generate", "rand" ]
